combine contents of data-preprocessing.py and data_preprocess.py
